main forecast file
This is the original script that created this entire forward momentum. This is when I realized that what shat I once thought couldn't be done by me actually could be done! This is the initial data scrap of surfcaptian.com for Oceanside California. (https://surfcaptain.com/forecast/oceanside-california) . along the way I exported it to an HTML file and added some css.